fix: TXE after public executor changes
